+++ allennlp/fairness/__init__.py
@@ -1,0 +1,12 @@
+from allennlp.fairness.bias_direction import (
+    PCABiasDirection,
+    PairedPCABiasDirection,
+    ClassificationNormalBiasDirection,
+    TwoMeansBiasDirection,
+)
+from allennlp.fairness.bias_mitigators import (
+    LinearBiasMitigator,
+    HardBiasMitigator,
+    INLPBiasMitigator,
+    OSCaRBiasMitigator,
+)